% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable-sync.R
\name{airtable_extract_photo}
\alias{airtable_extract_photo}
\title{Extract the first photo URL from an Airtable attachment field.}
\usage{
airtable_extract_photo(photo_field)
}
\description{
Extract the first photo URL from an Airtable attachment field.
}
\keyword{internal}
