% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/airtable-sync.R
\name{airtable_list_records}
\alias{airtable_list_records}
\title{List all records from an Airtable table, following pagination.}
\usage{
airtable_list_records(base_id, table, api_key)
}
\description{
List all records from an Airtable table, following pagination.
}
\keyword{internal}
