% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-utils.R
\name{announce_create_message}
\alias{announce_create_message}
\title{Create a formatted announcement message}
\usage{
announce_create_message(frontmatter, uri)
}
\arguments{
\item{frontmatter}{Named list from YAML front matter (needs \code{title},
\code{description}, \code{tags}).}

\item{uri}{URL of the post.}
}
\value{
Formatted message string.
}
\description{
Create a formatted announcement message
}
