% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-mastodon.R
\name{announce_post_mastodon}
\alias{announce_post_mastodon}
\title{Post an announcement to Mastodon}
\usage{
announce_post_mastodon(text, image = NULL, image_alt = NULL)
}
\arguments{
\item{text}{Post text.}

\item{image}{Optional file path to an image.}

\item{image_alt}{Alt text for the image.}
}
\value{
Post response (invisibly).
}
\description{
Post an announcement to Mastodon
}
