% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{at_scalar}
\alias{at_scalar}
\title{Read a scalar Airtable field as a length-1 character (NA if absent).}
\usage{
at_scalar(fields, key)
}
\description{
Read a scalar Airtable field as a length-1 character (NA if absent).
}
\keyword{internal}
