% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{at_vector}
\alias{at_vector}
\title{Read a multi-value Airtable field as a character vector.}
\usage{
at_vector(fields, key)
}
\description{
Read a multi-value Airtable field as a character vector.
}
\keyword{internal}
