% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/cf-ops.R
\name{cf_ops_format_workers_report}
\alias{cf_ops_format_workers_report}
\title{Format Workers invocation metrics as a markdown table}
\usage{
cf_ops_format_workers_report(df)
}
\arguments{
\item{df}{Data frame from \code{\link[=cf_ops_workers_invocations]{cf_ops_workers_invocations()}}.}
}
\value{
Character string with markdown-formatted report.
}
\description{
Format Workers invocation metrics as a markdown table
}
