% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/chapter-meetup-setup.R
\name{chapter_meetup_name}
\alias{chapter_meetup_name}
\title{The Meetup group name for a chapter}
\usage{
chapter_meetup_name(city)
}
\arguments{
\item{city}{Chapter city, used as written apart from trimming.}
}
\value{
The group name.
}
\description{
The Meetup group name for a chapter
}
