% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack-welcome.R
\name{common_welcome_channels}
\alias{common_welcome_channels}
\title{Channels common to both RLadies+ Slack workspaces}
\usage{
common_welcome_channels()
}
\value{
A list of named character vectors with \code{name} and \code{desc}.
}
\description{
Channels that exist in both the community and organisers Slack
workspaces, and are useful to surface to any new member regardless
of which workspace they joined.
}
