% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{contributor_thank}
\alias{contributor_thank}
\title{Thank a contributor when their PR is merged}
\usage{
contributor_thank(owner, repo, pr_number, author, org = "rladies")
}
\arguments{
\item{owner}{Repository owner.}

\item{repo}{Repository name.}

\item{pr_number}{PR number.}

\item{author}{GitHub login of the PR author.}

\item{org}{Organization name.}
}
\value{
Comment URL (invisibly).
}
\description{
Posts a thank-you message and adds the contributor to the repo's
contributor list if one exists.
}
