% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-pr.R
\name{directory_collision_issue}
\alias{directory_collision_issue}
\title{Flag a numeric-suffixed slug as a possible duplicate to merge.}
\usage{
directory_collision_issue(filename)
}
\description{
Flag a numeric-suffixed slug as a possible duplicate to merge.
}
\keyword{internal}
