% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_commit_message}
\alias{directory_commit_message}
\title{One-line commit message summarising the number of changed entries.}
\usage{
directory_commit_message(changes)
}
\description{
One-line commit message summarising the number of changed entries.
}
\keyword{internal}
