% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-pr.R
\name{directory_contact_social_issues}
\alias{directory_contact_social_issues}
\title{Flag contact methods that lack a matching \code{social_media} entry.}
\usage{
directory_contact_social_issues(entry)
}
\description{
Flag contact methods that lack a matching \code{social_media} entry.
}
\keyword{internal}
