% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_decode_json}
\alias{directory_decode_json}
\title{Decode a GitHub contents object into a parsed JSON list, or \code{NULL}.}
\usage{
directory_decode_json(obj)
}
\description{
Decode a GitHub contents object into a parsed JSON list, or \code{NULL}.
}
\keyword{internal}
