% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_entry_changes}
\alias{directory_entry_changes}
\title{Compute the pending file changes for a single submission.}
\usage{
directory_entry_changes(entry, org, repo, ref)
}
\description{
Compute the pending file changes for a single submission.
}
\keyword{internal}
