% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_entry_data}
\alias{directory_entry_data}
\title{Assemble the entry data list a submission contributes.}
\usage{
directory_entry_data(record, fields, lookups, slug)
}
\description{
Assemble the entry data list a submission contributes.
}
\keyword{internal}
