% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-pr.R
\name{directory_filename_issues}
\alias{directory_filename_issues}
\title{Filename convention issues, prefixed for the report.}
\usage{
directory_filename_issues(filename)
}
\description{
Filename convention issues, prefixed for the report.
}
\keyword{internal}
