% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_location}
\alias{directory_location}
\title{Build the \code{location} object, resolving the linked country.}
\usage{
directory_location(fields, country_lookup)
}
\description{
Build the \code{location} object, resolving the linked country.
}
\keyword{internal}
