% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_lookup}
\alias{directory_lookup}
\title{Map Airtable record ids to a labelling field.}
\usage{
directory_lookup(records, label_field)
}
\description{
Map Airtable record ids to a labelling field.
}
\keyword{internal}
