% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_merge_nested}
\alias{directory_merge_nested}
\title{Overlay a nested object's sub-keys, or replace wholesale if not both lists.}
\usage{
directory_merge_nested(target, val)
}
\description{
Overlay a nested object's sub-keys, or replace wholesale if not both lists.
}
\keyword{internal}
