% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_nested_keys}
\alias{directory_nested_keys}
\title{Entry keys that are objects and merged per sub-key rather than replaced.}
\usage{
directory_nested_keys()
}
\description{
Entry keys that are objects and merged per sub-key rather than replaced.
}
\keyword{internal}
