% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_overlay}
\alias{directory_overlay}
\title{Deep-merge \code{source} onto \code{target}, per-subkey for nested objects.}
\usage{
directory_overlay(target, source)
}
\description{
Deep-merge \code{source} onto \code{target}, per-subkey for nested objects.
}
\keyword{internal}
