% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_photo_meta}
\alias{directory_photo_meta}
\title{Extract photo download metadata (url, extension, credit) from a submission.}
\usage{
directory_photo_meta(fields)
}
\description{
Extract photo download metadata (url, extension, credit) from a submission.
}
\keyword{internal}
