% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_prefixed}
\alias{directory_prefixed}
\title{Collapse \verb{<prefix>*} scalar fields into a named list without the prefix.}
\usage{
directory_prefixed(fields, prefix)
}
\description{
Collapse \verb{<prefix>*} scalar fields into a named list without the prefix.
}
\keyword{internal}
