% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_r_groups}
\alias{directory_r_groups}
\title{Build the \code{r_groups} name -> url map from \verb{rgroup_name_*}/\verb{rgroup_url_*}.}
\usage{
directory_r_groups(fields)
}
\description{
Build the \code{r_groups} name -> url map from \verb{rgroup_name_*}/\verb{rgroup_url_*}.
}
\keyword{internal}
