% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-pr.R
\name{directory_review_entry}
\alias{directory_review_entry}
\title{Run all automated review checks on a single directory entry.}
\usage{
directory_review_entry(entry, filename, verify_handles = TRUE)
}
\description{
Run all automated review checks on a single directory entry.
}
\keyword{internal}
