% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_slug}
\alias{directory_slug}
\title{Slug from \code{directory_id}, falling back to \code{identifier}, cleaned to ASCII.}
\usage{
directory_slug(fields)
}
\description{
Slug from \code{directory_id}, falling back to \code{identifier}, cleaned to ASCII.
}
\keyword{internal}
