% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_social}
\alias{directory_social}
\title{Build the \code{social_media} object from \verb{some_*} fields, normalised.}
\usage{
directory_social(fields)
}
\description{
Build the \code{social_media} object from \verb{some_*} fields, normalised.
}
\keyword{internal}
