% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{directory_social_normalizers}
\alias{directory_social_normalizers}
\title{Field -> normaliser map for the \verb{some_*} social handles.}
\usage{
directory_social_normalizers()
}
\description{
Field -> normaliser map for the \verb{some_*} social handles.
}
\keyword{internal}
