% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_sort}
\alias{directory_sort}
\title{Recursively drop empty children and sort named lists for stable comparison.}
\usage{
directory_sort(x)
}
\description{
Recursively drop empty children and sort named lists for stable comparison.
}
\keyword{internal}
