% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_to_json}
\alias{directory_to_json}
\title{Serialise an entry to the directory's pretty JSON form.}
\usage{
directory_to_json(x)
}
\description{
Serialise an entry to the directory's pretty JSON form.
}
\keyword{internal}
