% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-manage.R
\name{empty_event_df}
\alias{empty_event_df}
\title{Create an empty events data frame}
\usage{
empty_event_df()
}
\value{
Data frame with 0 rows and standard event columns.
}
\description{
Create an empty events data frame
}
\keyword{internal}
