% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{escape_linkedin_chars}
\alias{escape_linkedin_chars}
\title{Escape special characters for LinkedIn API}
\usage{
escape_linkedin_chars(x)
}
\arguments{
\item{x}{Character string to escape.}
}
\value{
Escaped string.
}
\description{
Escape special characters for LinkedIn API
}
