% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-manage.R
\name{event_publish_summary}
\alias{event_publish_summary}
\title{Publish event summary as a GitHub issue}
\usage{
event_publish_summary(summary, org = "rladies", target_repo = "global-team")
}
\arguments{
\item{summary}{Formatted summary from \code{\link[=event_create_summary]{event_create_summary()}}.}

\item{org}{GitHub organization.}

\item{target_repo}{Repository to publish to.}
}
\value{
Issue URL (invisibly).
}
\description{
Publish event summary as a GitHub issue
}
