% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-awesome-creations.R
\name{format_awesome_item}
\alias{format_awesome_item}
\title{Dispatch an awesome-creations item to the package or content formatter}
\usage{
format_awesome_item(item, feed, src)
}
\description{
Dispatch an awesome-creations item to the package or content formatter
}
\keyword{internal}
