% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-format.R
\name{format_report_markdown}
\alias{format_report_markdown}
\title{Format a report as markdown}
\usage{
format_report_markdown(report)
}
\arguments{
\item{report}{Report data from \code{\link[=generate_report]{generate_report()}}.}
}
\value{
Character string with markdown-formatted report.
}
\description{
Format a report as markdown
}
