% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website-analytics.R
\name{format_website_analytics}
\alias{format_website_analytics}
\title{Format website analytics as markdown}
\usage{
format_website_analytics(analytics)
}
\arguments{
\item{analytics}{Data from \code{\link[=collect_website_analytics]{collect_website_analytics()}}.}
}
\value{
Formatted markdown string.
}
\description{
Format website analytics as markdown
}
