% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{gh_get_content}
\alias{gh_get_content}
\title{GET a repo file's contents object, or \code{NULL} if it does not exist.}
\usage{
gh_get_content(org, repo, path, ref = "main")
}
\description{
GET a repo file's contents object, or \code{NULL} if it does not exist.
}
\keyword{internal}
