% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-pkgdown-llms.R
\name{gh_list_org_r_repos}
\alias{gh_list_org_r_repos}
\title{List non-archived, non-disabled R-language repos in a GitHub org}
\usage{
gh_list_org_r_repos(org)
}
\description{
List non-archived, non-disabled R-language repos in a GitHub org
}
\keyword{internal}
