% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/templates.R
\name{gt_build_offboarding_body}
\alias{gt_build_offboarding_body}
\title{Build a global team offboarding issue body}
\usage{
gt_build_offboarding_body(team_slug, username, name)
}
\arguments{
\item{team_slug}{Team slug.}

\item{username}{GitHub username.}

\item{name}{Full name.}
}
\value{
Rendered issue body as a single character string.
}
\description{
Build a global team offboarding issue body
}
