% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-team-offboard.R
\name{gt_create_offboarding}
\alias{gt_create_offboarding}
\title{Create a global team offboarding issue}
\usage{
gt_create_offboarding(username, team, name = username, org = "rladies")
}
\arguments{
\item{username}{GitHub username.}

\item{team}{Team slug.}

\item{name}{Full name. Defaults to the username.}

\item{org}{GitHub organization. Defaults to \code{"rladies"}.}
}
\value{
The created issue URL (invisibly).
}
\description{
Create a global team offboarding issue
}
