% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-hugo-site.R
\name{html_node_to_markdown}
\alias{html_node_to_markdown}
\title{Convert an HTML node to GitHub-flavoured markdown via pandoc}
\usage{
html_node_to_markdown(node)
}
\description{
Convert an HTML node to GitHub-flavoured markdown via pandoc
}
\keyword{internal}
