% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{i18n_list_languages}
\alias{i18n_list_languages}
\title{List supported languages}
\usage{
i18n_list_languages()
}
\value{
Data frame with columns: code, name, native_name, direction.
}
\description{
List supported languages
}
