% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-hugo-site.R
\name{is_english_url}
\alias{is_english_url}
\title{Test whether a URL's first path segment is the English language root}
\usage{
is_english_url(url, src)
}
\description{
Test whether a URL's first path segment is the English language root
}
\keyword{internal}
