% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{li_get_version}
\alias{li_get_version}
\title{Get LinkedIn API version string}
\usage{
li_get_version()
}
\value{
Version string in YYYYMM format.
}
\description{
Get LinkedIn API version string
}
