% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{li_media_upload}
\alias{li_media_upload}
\title{Upload media to LinkedIn}
\usage{
li_media_upload(author, media)
}
\arguments{
\item{author}{LinkedIn URN of the owner.}

\item{media}{File path to the image.}
}
\value{
LinkedIn image ID string.
}
\description{
Upload media to LinkedIn
}
