% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{li_oauth}
\alias{li_oauth}
\title{Perform LinkedIn OAuth authentication}
\usage{
li_oauth()
}
\value{
OAuth token object.
}
\description{
Interactive OAuth2 flow for obtaining a LinkedIn token.
}
