% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{li_post_write}
\alias{li_post_write}
\title{Post to LinkedIn}
\usage{
li_post_write(author, text, image = NULL, image_alt = "")
}
\arguments{
\item{author}{LinkedIn URN of the author.}

\item{text}{Post text.}

\item{image}{Optional file path to an image.}

\item{image_alt}{Alt text for the image.}
}
\value{
LinkedIn post ID (invisibly).
}
\description{
Post to LinkedIn
}
