% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{li_req}
\alias{li_req}
\title{Create a base LinkedIn API request}
\usage{
li_req(endpoint_version = "rest", ...)
}
\arguments{
\item{endpoint_version}{API version path. Defaults to \code{"rest"}.}

\item{...}{Passed to \code{li_req_auth()}.}
}
\value{
Configured httr2 request.
}
\description{
Create a base LinkedIn API request
}
