% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-linkedin.R
\name{li_urn_me}
\alias{li_urn_me}
\title{Get the LinkedIn URN for the authenticated user}
\usage{
li_urn_me()
}
\value{
URN string like \code{"urn:li:person:abc123"}.
}
\description{
Get the LinkedIn URN for the authenticated user
}
