% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/i18n.R
\name{list_supported_languages}
\alias{list_supported_languages}
\title{List supported languages}
\usage{
list_supported_languages()
}
\value{
Data frame with columns: code, name, native_name, direction.
}
\description{
List supported languages
}
