% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-manage.R
\name{load_events_config}
\alias{load_events_config}
\title{Load events configuration}
\usage{
load_events_config()
}
\value{
Named list from \code{inst/config/events.yml}.
}
\description{
Load events configuration
}
\keyword{internal}
