% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_labels_config}
\alias{load_labels_config}
\title{Load file-path to label mappings}
\usage{
load_labels_config()
}
\value{
Named list with label mappings.
}
\description{
Load file-path to label mappings
}
