% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{load_teams_config}
\alias{load_teams_config}
\title{Load teams configuration}
\usage{
load_teams_config()
}
\value{
Named list with organization, global_team_id, teams, and
default_assignees.
}
\description{
Load teams configuration
}
