% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-hugo-site.R
\name{normalise_hugo_url}
\alias{normalise_hugo_url}
\title{Normalise a Hugo URL by stripping a trailing /index.html}
\usage{
normalise_hugo_url(url)
}
\description{
Normalise a Hugo URL by stripping a trailing /index.html
}
\keyword{internal}
