% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{normalize_bluesky}
\alias{normalize_bluesky}
\title{Normalise a Bluesky handle to a bare lowercase handle.}
\usage{
normalize_bluesky(x)
}
\description{
Normalise a Bluesky handle to a bare lowercase handle.
}
\keyword{internal}
