% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{normalize_github}
\alias{normalize_github}
\title{Normalise a GitHub handle to a bare lowercase username.}
\usage{
normalize_github(x)
}
\description{
Normalise a GitHub handle to a bare lowercase username.
}
\keyword{internal}
