% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{normalize_linkedin}
\alias{normalize_linkedin}
\title{Normalise a LinkedIn handle to a bare lowercase username.}
\usage{
normalize_linkedin(x)
}
\description{
Normalise a LinkedIn handle to a bare lowercase username.
}
\keyword{internal}
