% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{normalize_mastodon}
\alias{normalize_mastodon}
\title{Normalise a Mastodon profile URL to \verb{@user@instance} handle form.}
\usage{
normalize_mastodon(x)
}
\description{
Normalise a Mastodon profile URL to \verb{@user@instance} handle form.
}
\keyword{internal}
