% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{normalize_twitter}
\alias{normalize_twitter}
\title{Normalise a Twitter/X handle to a bare lowercase username.}
\usage{
normalize_twitter(x)
}
\description{
Normalise a Twitter/X handle to a bare lowercase username.
}
\keyword{internal}
