% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-bluesky.R
\name{post_bluesky}
\alias{post_bluesky}
\title{Post an announcement to Bluesky}
\usage{
post_bluesky(text, image = NULL, image_alt = NULL)
}
\arguments{
\item{text}{Post text.}

\item{image}{Optional file path to an image.}

\item{image_alt}{Alt text for the image.}
}
\value{
Post response (invisibly).
}
\description{
Post an announcement to Bluesky
}
