% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/analytics-publish.R
\name{publish_analytics_dashboard}
\alias{publish_analytics_dashboard}
\title{Publish analytics dashboard as a GitHub issue}
\usage{
publish_analytics_dashboard(
  dashboard_data,
  org = "rladies",
  target_repo = "global-team"
)
}
\arguments{
\item{dashboard_data}{Data from \code{\link[=generate_analytics_dashboard]{generate_analytics_dashboard()}}.}

\item{org}{GitHub organization.}

\item{target_repo}{Repository to publish to.}
}
\value{
Issue URL (invisibly).
}
\description{
Publish analytics dashboard as a GitHub issue
}
