% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-activity.R
\name{report_publish}
\alias{report_publish}
\title{Publish a report as a GitHub issue}
\usage{
report_publish(report, target_repo = "global-team", org = "rladies")
}
\arguments{
\item{report}{Report data from \code{\link[=report_generate]{report_generate()}}.}

\item{target_repo}{Repository to publish to. Defaults to \code{"global-team"}.}

\item{org}{Organization. Defaults to \code{"rladies"}.}
}
\value{
Issue URL (invisibly).
}
\description{
Publish a report as a GitHub issue
}
