% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{punct_name}
\alias{punct_name}
\title{Append a period to single-character words in a name.}
\usage{
punct_name(x)
}
\description{
Append a period to single-character words in a name.
}
\keyword{internal}
