% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-challenge.R
\name{r_challenge_difficulty_badge}
\alias{r_challenge_difficulty_badge}
\title{Slack emoji badge for a difficulty level}
\usage{
r_challenge_difficulty_badge(difficulty)
}
\arguments{
\item{difficulty}{One of the levels in \code{\link[=r_challenge_difficulties]{r_challenge_difficulties()}}.}
}
\value{
Character scalar emoji.
}
\description{
Slack emoji badge for a difficulty level
}
