% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-challenge.R
\name{r_challenge_difficulty_label}
\alias{r_challenge_difficulty_label}
\title{Human-readable label for a difficulty level}
\usage{
r_challenge_difficulty_label(difficulty)
}
\arguments{
\item{difficulty}{One of the levels in \code{\link[=r_challenge_difficulties]{r_challenge_difficulties()}}.}
}
\value{
Character scalar label.
}
\description{
Human-readable label for a difficulty level
}
