% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/r-challenge.R
\name{r_challenge_issue_title}
\alias{r_challenge_issue_title}
\title{GitHub issue title for a drafted challenge}
\usage{
r_challenge_issue_title(challenge)
}
\arguments{
\item{challenge}{Parsed challenge from \code{\link[=r_challenge_parse]{r_challenge_parse()}}.}
}
\value{
Character scalar issue title.
}
\description{
GitHub issue title for a drafted challenge
}
