% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-github-remote-files.R
\name{remote_file_chunks}
\alias{remote_file_chunks}
\title{Fetch a file's raw contents from the GitHub API and chunk the markdown}
\usage{
remote_file_chunks(entry, src, token)
}
\description{
Fetch a file's raw contents from the GitHub API and chunk the markdown
}
\keyword{internal}
