% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-github-org.R
\name{render_team_text}
\alias{render_team_text}
\title{Render a GitHub team to a labelled text block}
\usage{
render_team_text(team)
}
\description{
Render a GitHub team to a labelled text block
}
\keyword{internal}
