% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-github-org.R
\name{repo_to_chunks}
\alias{repo_to_chunks}
\title{Render a GitHub repo to metadata and README chunks}
\usage{
repo_to_chunks(repo, token)
}
\description{
Render a GitHub repo to metadata and README chunks
}
\keyword{internal}
