% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/report-format.R
\name{report_format_markdown}
\alias{report_format_markdown}
\title{Format a report as markdown}
\usage{
report_format_markdown(report)
}
\arguments{
\item{report}{Report data from \code{\link[=report_generate]{report_generate()}}.}
}
\value{
Character string with markdown-formatted report.
}
\description{
Format a report as markdown
}
