% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable.R
\name{resolve_links}
\alias{resolve_links}
\title{Resolve a vector of Airtable link ids to their looked-up labels.}
\usage{
resolve_links(ids, lookup)
}
\description{
Resolve a vector of Airtable link ids to their looked-up labels.
}
\keyword{internal}
