% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website-rum.R
\name{rum_format_analytics}
\alias{rum_format_analytics}
\title{Format Cloudflare RUM analytics as markdown}
\usage{
rum_format_analytics(analytics)
}
\arguments{
\item{analytics}{Data from \code{\link[=rum_collect_analytics]{rum_collect_analytics()}}.}
}
\value{
Character string with markdown-formatted report.
}
\description{
Format Cloudflare RUM analytics as markdown
}
