% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-newsletter.R
\name{send_newsletter}
\alias{send_newsletter}
\title{Send a newsletter via ConvertKit}
\usage{
send_newsletter(frontmatter, url)
}
\arguments{
\item{frontmatter}{Named list with \code{seo}, \code{image}, \code{title} fields.}

\item{url}{URL of the post.}
}
\value{
API response (invisibly).
}
\description{
Send a newsletter via ConvertKit
}
