% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-utils.R
\name{short_url}
\alias{short_url}
\title{Shorten a URL using the Short.io API}
\usage{
short_url(uri)
}
\arguments{
\item{uri}{URL to shorten.}
}
\value{
Shortened URL string.
}
\description{
Shorten a URL using the Short.io API
}
