% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/welcome.R
\name{slack_conversations_open}
\alias{slack_conversations_open}
\title{Open a Slack direct message channel}
\usage{
slack_conversations_open(team_id, user_id, workspace)
}
\arguments{
\item{team_id}{Slack team id.}

\item{user_id}{Slack user id to DM.}

\item{workspace}{\code{"organiser"} or \code{"community"}.}
}
\value{
The DM channel ID, or \code{NULL} on failure.
}
\description{
Open a Slack direct message channel
}
