% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-hugo-site.R
\name{strip_suffix}
\alias{strip_suffix}
\title{Strip a trailing suffix from a string if present}
\usage{
strip_suffix(s, suffix)
}
\description{
Strip a trailing suffix from a string if present
}
\keyword{internal}
