% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/announce-utils.R
\name{tags2hash}
\alias{tags2hash}
\title{Convert tags to hashtag string}
\usage{
tags2hash(tags)
}
\arguments{
\item{tags}{Character vector of tags.}
}
\value{
Single string with hashtags.
}
\description{
Convert tags to hashtag string
}
