% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/config.R
\name{team_slugs}
\alias{team_slugs}
\title{List all valid team slugs}
\usage{
team_slugs(config = NULL)
}
\arguments{
\item{config}{Teams config as returned by \code{\link[=load_teams_config]{load_teams_config()}}. Loaded
automatically if \code{NULL}.}
}
\value{
Character vector of team slugs.
}
\description{
List all valid team slugs
}
