% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{update_contributors_list}
\alias{update_contributors_list}
\title{Generate and update a contributors list for a repo}
\usage{
update_contributors_list(
  owner,
  repo,
  file_path = ".github/contributors.md",
  format = "grid"
)
}
\arguments{
\item{owner}{Repository owner.}

\item{repo}{Repository name.}

\item{file_path}{Path in the repo to update with the contributor list.
Defaults to \code{".github/contributors.md"}.}

\item{format}{Format for the contributor list.}
}
\value{
PR URL if changes were made, \code{NULL} otherwise (invisibly).
}
\description{
Fetches contributors and creates/updates a PR with the contributor
list in a specified file.
}
