% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/rag-source-hugo-site.R
\name{url_path}
\alias{url_path}
\title{Return the path component of a URL, defaulting to "/"}
\usage{
url_path(url)
}
\description{
Return the path component of a URL, defaulting to "/"
}
\keyword{internal}
