% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website-analytics.R
\name{website_format_analytics}
\alias{website_format_analytics}
\title{Format website analytics as markdown}
\usage{
website_format_analytics(analytics)
}
\arguments{
\item{analytics}{Data from \code{\link[=website_collect_analytics]{website_collect_analytics()}}.}
}
\value{
Formatted markdown string.
}
\description{
Format website analytics as markdown
}
