% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/website-analytics.R
\name{website_publish_report}
\alias{website_publish_report}
\title{Publish website analytics as a GitHub issue}
\usage{
website_publish_report(
  report_data,
  org = "rladies",
  target_repo = "global-team",
  slack_channel = NULL
)
}
\arguments{
\item{report_data}{Data from \code{\link[=website_generate_report]{website_generate_report()}}.}

\item{org}{GitHub organization.}

\item{target_repo}{Repository to publish to.}

\item{slack_channel}{Optional Slack channel to post a summary to.}
}
\value{
Issue URL (invisibly).
}
\description{
Publish website analytics as a GitHub issue
}
