% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/contributors.R
\name{contributor_welcome}
\alias{contributor_welcome}
\title{Welcome a first-time contributor}
\usage{
contributor_welcome(owner, repo, number, author, is_pr = TRUE, org = "rladies")
}
\arguments{
\item{owner}{Repository owner.}

\item{repo}{Repository name.}

\item{number}{Issue or PR number.}

\item{author}{GitHub login of the author.}

\item{is_pr}{Whether this is a PR (TRUE) or issue (FALSE).}

\item{org}{Organization name.}
}
\value{
Comment URL or \code{NULL} if author is a team member (invisibly).
}
\description{
Checks if the PR/issue author has previous contributions to the repo.
If this is their first, posts a warm welcome message. If not first-time,
posts a shorter thank-you.
}
