% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-airtable-io.R
\name{directory_create_pr}
\alias{directory_create_pr}
\title{Commit recorded changes to the sync branch and open (or reuse) a PR.}
\usage{
directory_create_pr(changes, deletes, org, repo, base = "main")
}
\description{
Commit recorded changes to the sync branch and open (or reuse) a PR.
}
\keyword{internal}
