% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/directory-pr.R
\name{directory_review_format}
\alias{directory_review_format}
\title{Format the per-entry review results as a markdown PR comment.}
\usage{
directory_review_format(reviews)
}
\description{
Format the per-entry review results as a markdown PR comment.
}
\keyword{internal}
