% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/gh.R
\name{gh_thank_contributor}
\alias{gh_thank_contributor}
\title{Thank a contributor when their PR is merged}
\usage{
gh_thank_contributor(owner, repo, pr_number, author, org = "rladies")
}
\arguments{
\item{owner}{Repository owner.}

\item{repo}{Repository name.}

\item{pr_number}{PR number.}

\item{author}{GitHub login of the PR author.}

\item{org}{Organization name.}
}
\value{
Comment URL (invisibly).
}
\description{
Posts a thank-you message tailored to whether this is their first
merged PR.
}
