% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/global-team-offboard.R
\name{global_team_finalize_offboarding}
\alias{global_team_finalize_offboarding}
\title{Finalize global team offboarding by removing user from teams}
\usage{
global_team_finalize_offboarding(username, team, org = "rladies")
}
\arguments{
\item{username}{GitHub username.}

\item{team}{Team slug.}

\item{org}{GitHub organization. Defaults to \code{"rladies"}.}
}
\value{
Invisibly returns \code{NULL}. Called for its side effect of removing
the user from the specified teams.
}
\description{
Finalize global team offboarding by removing user from teams
}
