% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/event-manage.R
\name{publish_event_summary}
\alias{publish_event_summary}
\title{Publish event summary as a GitHub issue}
\usage{
publish_event_summary(summary, org = "rladies", target_repo = "global-team")
}
\arguments{
\item{summary}{Formatted summary from \code{\link[=create_event_summary]{create_event_summary()}}.}

\item{org}{GitHub organization.}

\item{target_repo}{Repository to publish to.}
}
\value{
Issue URL (invisibly).
}
\description{
Publish event summary as a GitHub issue
}
