% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/slack-manage.R
\name{slack_subscribe_rss}
\alias{slack_subscribe_rss}
\title{Subscribe an RSS feed to a Slack channel}
\usage{
slack_subscribe_rss(
  rss_url,
  channel = "rladiesblogs",
  token = Sys.getenv("SLACK_TOKEN")
)
}
\arguments{
\item{rss_url}{RSS feed URL to subscribe.}

\item{channel}{Slack channel name (without #). Defaults to "rladiesblogs".}

\item{token}{Slack API token.}
}
\value{
API response (invisibly).
}
\description{
Posts a \verb{/feed subscribe} command to the specified Slack channel
using the Slack API.
}
